Create structure for visual half of webpage
The left half of the webpage contains a background image and a
company logo and name to capture the user's attention. The
structure for this part of the webpage has been created, but it
still needs to be styled. Additionally, the name of the background
image file has been changed to something more memorable.