Finish styling of visual half of webpage
The left half of the webpage already included a custom background
image and some basic styling. However, more styling has been added.
This inlcudes adding a dark background for the webpage logo and name
to make it stand out visually, changing the font color of the credit
at the bottom so that it contrasts more with the background. and
clamping the width of the image between two values so that it does
not become too big or too smal depending on the page size and
changes dynamically as the page is resized.