Change spacing and font sizes in form
The form was cluttered before. Now, the more important p elements
at the top of the page have been given a bigger font size, the form
element labels have been given a smaller font size, and the spacing
between elements has been increased on the right half of the webpage.