Style form elements
The "Log in" link still had default styling. It has been changed
to match the color of the submit button. Additionally, in the case
of the passwords not matching, there needs to be a message that
tells the user that the passwords do not match. This has been added,
given a smaller font size, made red, and has been set to hidden by
default.