Create error message for matching password fields
An error message needs to be displayed if the two password fields
do not match. Before, this was accomplished with the ::after pseudo-
element, but this cannot be changed dynamically, so it did not work.
Now, there is a new span element for this error message. Its
visibility will be updated dynamically when the passwords do and do
not match.