Add password checking functionality
When the user clicks "Create Account", the password and confirm
password fields need to be checked to ensure they match before the
form is submitted. This has been accomplished with JavaScript. If
the use submits the form and the password fields do not match, then
the form is not submitted and an error message is displayed. the
error message goes away once the user starts typing in one of the
password fields.